restructure to include global state
